% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bovs2.r
\name{baggingTest}
\alias{baggingTest}
\title{baggingTest function}
\usage{
baggingTest(res, model_list, x_test, y_test, iteration = 1000)
}
\arguments{
\item{res}{A vector obtained by the baggingVal function,  the length is equal to the number of bootstrap done by the baggingVal function}

\item{model_list}{The list is same as the one used in baggingVal function to produce res}

\item{x_test}{A dataframe contains the features used by the bagginTest function}

\item{y_test}{A vector contains the labels used by baggingTest function}

\item{iteration}{Total number of iterations needed}
}
\description{
This function allows you to make use of the training result obtained by bagginVal to test the performance of new model
}
