% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bovs2.r
\name{baggingVal}
\alias{baggingVal}
\title{baggingVal function}
\usage{
baggingVal(model_list, x_val, y_val, iteration = 1000)
}
\arguments{
\item{x_val}{a dataframe of features to be used as validation set}

\item{y_val}{a vector indiates what is the label}

\item{iteration}{the total number of bootstrap to be used, the default value is 1000}

\item{model}{the list of deep learning models returned by model in keras R}
}
\description{
This function allows you to do bagging on the validation set
}
\examples{
baggingVal(model_list, x_val, y_val)
}
