% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist-data.r
\docType{data}
\name{mnist}
\alias{mnist}
\title{MNIST database (Modified National Institute of Standards and Technology database)}
\format{An object of class \code{list} of length 2.}
\usage{
data(mnist)
}
\description{
A large database of handwritten digits that is commonly used for training various image processing systems.
}
\examples{
data(mnist)
}
\keyword{datasets}
