% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bovs2.r
\name{performanceEvaluate}
\alias{performanceEvaluate}
\title{Performance evaluation function}
\usage{
performanceEvaluate(model_list, x_test, y_test)
}
\arguments{
\item{model_list}{A list of neural network models object returned by keras R}

\item{x_test}{A dataframe contains the features used by the performanceEvaluate function}

\item{y_test}{A vector contains the labels used by baggingTest function}
}
\description{
This function is used to evaluate the performance of a series of models trained by R keras
}
