% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bovs2.r
\name{sortMaxFrequency}
\alias{sortMaxFrequency}
\title{sortMaxFrquency function}
\usage{
sortMaxFrequency(frequencyVec)
}
\arguments{
\item{frequencyVec}{A vector contains the fequency of the appearance of the element}
}
\description{
This function allows to find the item that has the maximum frequency of appearance in a vector
}
